Basic pages: create folder & delete support, and "go up" to navigate up in folders (#140)
* WIP implement DELETE

* Add create folder, delete, and "go up" support for basic pages

* Add tests for new functionality

+++ bulgur-cloud-backend/src/storage.rs
@@ -1,4 +1,5 @@
 use std::{
+    cmp::Ordering,
     io,
     ops::Deref,
     path::{Path, PathBuf},
@@ -162,6 +163,18 @@ pub async fn get_storage_internal(
                 size: meta.len(),
             })
         }
+        // Sort them folders first, then files. Sorted by name within these
+        // groups. The react UI does it's own sorting internally on top of this,
+        // but this is helpful for the basic UI.
+        folder_contents.sort_by(|a, b| {
+            if !a.is_file && b.is_file {
+                return Ordering::Less;
+            }
+            if !b.is_file && a.is_file {
+                return Ordering::Greater;
+            }
+            return a.name.cmp(&b.name);
+        });
         Ok(Either::Right(web::Json(FolderResults {
             entries: folder_contents,
         })))
@@ -189,15 +202,40 @@ async fn delete_storage(
 ) -> Result<HttpResponse, StorageError> {
     let (store, path) = params.as_ref();
 
-    let store_path = get_authorized_path(&authorized, store, Some(path))?;
-    if fs::metadata(&store_path).await?.is_file() {
-        tracing::debug!("Deleting file {:?}", store_path);
-        fs::remove_file(&store_path).await?;
-    } else {
-        tracing::debug!("Deleting folder {:?}", store_path);
-        fs::remove_dir_all(&store_path).await?;
-    }
+    common_delete(&authorized, store, Some(path)).await?;
     Ok(empty_ok_response())
+}
+
+/// Checks that the user is authorized to delete this path, and then deletes it.
+///
+/// Returns the deleted path.
+pub async fn common_delete(
+    authorized: &Option<ReqData<Authorized>>,
+    store: &str,
+    path: Option<&str>,
+) -> Result<PathBuf, StorageError> {
+    let store_path = get_authorized_path(&authorized, store, path)?;
+    match path {
+        Some(path) => {
+            if path.is_empty() {
+                // Don't allow empty paths, otherwise the store itself will get
+                // deleted
+                Err(StorageError::BadPath)
+            } else {
+                if fs::metadata(&store_path).await?.is_file() {
+                    tracing::debug!("Deleting file {:?}", store_path);
+                    fs::remove_file(&store_path).await?;
+                } else {
+                    tracing::debug!("Deleting folder {:?}", store_path);
+                    fs::remove_dir_all(&store_path).await?;
+                }
+                Ok(store_path)
+            }
+        }
+        // Don't allow missing paths, otherwise the store itself will get
+        // deleted
+        None => Err(StorageError::BadPath),
+    }
 }
 
 #[tracing::instrument]
